transaction-service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
kafka-python==2.0.2
redis==4.3.4
uuid==1.30
requests==2.28.1
flask-cors==3.0.10 